users can now rename groups
